pyproject.toml
[tool.poetry]
name = "todo-practice"
version = "0.1.0"
description = ""
authors = ["Kosuke Sugiura <koosuke711@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
